Use TaskContainer.actualize() (renamed to 'realize') instead of similar method on Project.
This is largely just to not have two different ways of doing the same thing.
There is a semantic change in here that the new method traverses dependencies, which may bring more tasks into existence when task dependencies are expressed in some indirect way that triggers a task to be created when resolving task dependencies.

+review REVIEW-5370