Define bridged collection items in terms of retrieval from the collection's model node, instead of the collection itself.
This is generally about reducing implicit state with our bridging, for the task container in particular.
Implicit state (i.e. state dragged in by function closure) is problematic when resetting ephemeral state.
References to epehemeral elements (e.g. the task container) closed over in item creators prevent the items being created fresh the next time around.

An important change here is that each bridged collection now has two implicit “invisible” nodes.
One for the backing store and one for the instantiator.
These are for item creators to depend on, in order to be defined in terms of the model node instead of the actual container.

+review REVIEW-5382