Documented need to check IncrementalTaskInputs.incremental
I use IncrementalTaskInputs a lot, but until recently did never check for
inputs.incremental to cleanup before doing any work. This resulted in stale
files still laying around in the output dir, because I thought
inputs.removed() would take care of this. This is wrong. The documentation
just did not state clearly enough this is necessary.