Add BuildOperationLogger and factory
A task will start() and overall build operation, execute the
underlying operations and then call done() to wrap up the overall
action.

For the native tools, we currently buffer the output in
DefaultCommandLineToolInvocationWorker and only try to write the
output once the operation is complete/failed.

The BuildOperationLogger writes the output for the operations to
a log file and writes up to N-operations to Gradle's standard out.

+review REVIEW-5406