Fix the formatting auto config for the latest IDEA version.
I've moved the control of indents to a .editorconfig file, which IDEA supports through a plugin (that the user is prompted to install when opening the project). This ensures that at least the basic whitespace settings continue to work, and now also are respected in all editors that support .editorconfig.