Simplify implementation of triggering a build in continuous mode.
Removes the formal 'trigger' abstraction, replacing with lower level primitives directly the build action executer. The primary benefit here is fewer moving parts and types.

As invoking the listening is now done at the end of the build execution loop instead of via a build listener, the task inputs accumulator had to be elevated to global scope. This is something to be addressed in forthcoming commits.

+review REVIEW-5481