Introduce BuildExecuter as a marker subclass of BuildActionExecuter<BuildActionParameters> that demarcates the “real” build executer.
The thing that implements this is wrapped in a bunch of places to support different execution modes and as such is depended up on the service registry wirings. Previously, such decorators depended on the concrete InProcessBuildActionExecuter concrete type that happened to, but no longer does, do what BuildExecuter now signifies. The base contract type BuildActionExecuter<BuildActionParameters> cannot be used to extract this because there are multiple implementations of this type.

+review REVIEW-5459