Don't assume that the instantiator for a domain object collection backed model map implicitly adds to the container.
Also, don't wrap the componentSpec.binaries model map in another model map implementation. This is a slight functional regression in that the ModelMap of binaries given to @ComponentBinaries rule no longer prevents mutation after the method has executed, in the same way that componentSpec.binaries does not (and a lot of other collections/objects). This is an acceptable trade off, for now, in order to be able to remove another ModelMap implementation. This problem will have to be solved pervasively later.

+review REVIEW-5481