Implement a quiet period for continuous mode.
This is needed now to make the integration tests reliable. The JDK watcher service fires events immediately on the creation of new files, before the actual content is written. Tests were failing as the state of files at build time didn't match expectations, causing different build results.

This implementation triggers the build after 250ms of inactivity after the first change. This is not configurable.