Don't register the backing dir of MapFileTree as something to watch for changes.
This dir will contain content that is generated when the file tree is visited, by executing code. That is, it's not going to contain user editable files.

In theory, we should be detecting when an action is some kind of transform on an actual user editable file, but we don't have that kind of insight into what the functions do.

+review REVIEW-5511