Document existing behavior for input/output task dependencies
This documents the existing behavior of the @Input* and @Output* task
dependencies. In particular it clarifies that how it impacts the
"up-to-date" checks on each task... and whether it depends on the content
of a annotated file or directory.

It also explains how to configure a dependency on only the file or
directory path and not the contents of it.