Fix graph traversal with complex cyclic graphs
This bug revealed itself as a NPE running 'gradle tasks' on the Gradle build.
By isolating the minimal failing case and reproducing in a unit test, I found
the issue: we weren't setting the root node for all nodes in a cycle, meaning that the incorrect
cycle root could be used when a cycle-causing node was added after an inner cycle was already resolved.