GRADLE-2316 Detect TCP self-connect and retry connection
When attempting to connect to the daemon, if the daemon has not bound
to the port, there is a small chance that the OS may attempt to bind and
connect to the same port.  When this happens, data written to this socket
is "reflected" back on the receiving side.

This change attempts to detect that by comparing the local and remote
addresses.  If they are the same, we assume that we have self-connected
and need to retry the connection.  Since ports are assigned sequentially
most of the time, this should work on the second retry.

+review REVIEW-5610