Normalize line wrapping in jdk9 design doc
This change eliminates arbitrary intra-paragraph newlines in favor of
keeping paragraphs on a single line. Doing so has several benefits:

 - Facilitates the effective use of soft wrap settings in the user's
   editor of choice

 - Facilitates the effective use of git's `--word-diff` switch at the
   command line and/or GitHub's `?w1` query string parameter, so as to
   avoid seeing spurious whole-line diffs due only to the introduction
   of a newline

 - Better aligns the markdown source to the HTML generated by GitHub
   Flavored Markdown (GFM), which ignores intra-paragraph newlines, and
   flows lines together as if they were one to begin.

There does not appear to be a strong convention with regard to line
length and wrapping throughout the design-docs directory. This commit
reflects a general recommendation that we favor not wrapping from now
on.