Remove redundant assertions in integration tests
This change removes use of Groovy's assert keyword in statements where a
JUnit-style `assert*` method is already being invoked. It is assumed
that these were introduced by mistake, given that there are only several
such instances throughout the codebase, and that the assert keyword
could have no effect anyway (because any such statement will be
short-circuited when the `assert*` method throws an exception).