Introduce JVM library API specification DSL
This change allows JVM library authors to specify and constrain the set
of packages that constitute a library's public API, e.g.:

    model {
        components {
            main(JvmLibrarySpec) {
                api {
                    exports 'com.example.p1'
                    exports 'com.example.p2'
                }
            }
        }
    }

The `api { }` DSL is backed by the newly-introduced `ApiSpec`, and
arguments to its `exports` clause are validated against JLS rules
governing package names. See Javadoc of the newly-introduced
`PackageName` class for details.

Aside from other basic validation, the presence (or absence) of an api
specification does not yet affect the output of the build; this will be
accomplished in subsequent commits. When implemented, it is assumed that
a `#getApiSpec` method will be exposed by the `JvmLibrarySpec`
interface, and that it will be called within
`JvmComponentPlugin#createBinaries` in order to generate 'api' and
'impl' jars that respect the api specification.

Basic user guide documentation for this feature is also introduced,
though it is essentially a placeholder at this point. As the overall
feature progresses toward completion, this section of the user guide
should be updated accordingly.

DSL documentation is *not* included here, as there is currently little
or no coverage of any software model components in the DSL docs. It is
assumed that a dedicated effort will be required to get the DSL docs up
to speed with the software model, and that api specification DSL
documentation will be added at that time.

Implementation notes and questions:

 - On visibility:

 The newly-introduced `ApiSpec` and `PackageName` types are
 package-private and concrete, i.e. do not conform to the
 <<InterfaceName>> and Default<<InterfaceName>> convention typical
 throughout the Gradle codebase. This has been done intentionally so as
 to encourage a discussion on these topics during code review. Likewise
 for the newly-introduced and intentionally package-private
 `DefaultJvmLibrarySpec#api` method.

 - On validation:

 This change implements api spec validation in a "fail-fast" manner,
 i.e., validation occurrs immediately on processing each `exports`
 clause, and any validation error results in the immediate throwing of a
 `InvalidUserDataException`. The software model offers first-class
 support for validation of model elements in the form of
 `@Validate`-annotated methods, but experiments in validating api specs
 in this manner proved awkward and verbose. It should be considered
 during code review whether `@Validate`-style deferred validation is
 indeed important and/or necessary, and if so, how best to do it.

 - On the use of Javadoc for internal and/or incubating types:

 Newly-introduced internal types have been documented with Javadoc; this
 is a departure from the norm, but again done intentionally so as to
 spur discussion on the topic.

+review REVIEW-5648