Allow binary names to reflect their relationship to their containing component, rather than their position in the space.
Previously, a binary's `name` property was expected to be defined in such a way so that it was unique across all binaries in the same project. This lead to various naming schemes for binaries that encoded their owning component's identity into the name. Now, a binary's `name` is required to be unique only amongst the binaries of the same component. This means that the binary's name can now reflect its relationship to the component.

An encoding is applied to binary names when using the name in some other namespace, such as when the binary is added to the `binaries` container, or when the 'build' and language transform tasks are defined for the binary.

This is a breaking change: the binary will appear using a different name under `binaries`, and its tasks will use different names under `tasks`.