Changed the JVM and native component plugins to use names for binaries that reflect the role they play.
Now that more than one component may have a binary with a given name, use:

-  `jar` as the name for a Jar binary that belongs to a JVM library
- `executable` as the name for an executable that belongs to a native application
-  `sharedLibrary` and `staticLibrary` as the name for the binaries of a native library.

These can later turn into static properties of the appropriate `ComponentSpec` types.

There is a breaking change in this commit: the lifecycle task name and output directory name of the binaries of components with multiple variants now include the component name at the start of the name, rather than somewhere in the middle.