Added workaround so that binaries appear in the `binaries` container as their public type, and re-enabled some disabled tests.
This is a workaround, while waiting for further changes. With this workaround, a binary is visible only as its public type, plus some, but not all, of the built-in internal view types. All other views are lost. This workaround also discards other details of the binary, including state transitions (but this is not a new problem).

A better solution would add a reference to the binary node into the `binaries` container, preserving all information about the node, including its views, state transitions and so on. This would require `BaseBinarySpec` to be node backed. Alternatively, we might solve the various use cases in different ways that don't require adding the binaries to a top-level container.