Use a real `ModelMap` for `ComponentSpec.sources`
This commit removes the backing `FunctionalSourceSet` instance from `BaseComponentSpec`,
instead using a real `ModelMap` implementation backed by a `ModelNode`. This removes
a lot of complexity from the implementation of `BaseComponentSpec`, as well as
making the behaviour of `ComponentSpec.sources` consistent with other `ModelMap`
properties.

Each `LanguageSourceSet` requires a 'parentName' property to be set, which is used
for providing a project-scoped unique name as well as source default locations.
For an element in `ComponentSpec.sources`, the 'parentName' is equal to the owning
component name. To make this possible, `LanguageSourceSetNodeIntializer` is now
aware of this relationship, and will set the 'parentName' appropriately. This is
somewhat consistent with the way the `owner` component is set when constructing
a `BaseBinarySpec` instance.

Notes:
- Elements of `ComponentSpec.sources` are now configured on demand.
- Deleted tests that verified sources sets couldn't be removed: these are
  no longer relevant with a `ModelMap` implementation.

Issue: gradle/langos#40 +review REVIEW-5681