Move o.g.jvm.{internal.apigen.abi=>tasks.api}
This commit completes the process of consolidating the internal.apigen
and internal.apigen.abi packages into a single new, (and non-internal)
package: o.g.jvm.tasks.api.

The rationale for this is that these types all work together toward a
single purpose. That is, they are highly cohesive in working together to
produce an API Jar. This means that there is not only no need to split
them across two packages, there actually an important reason *not* to do
so. What's the reason? Find out in the next commit.