Use a custom node initializer strategy for empty subtypes of `ModelMap`
This change allows us to define subtypes of `ModelMap` and create instances
using a standard `@Model` rule. Previously it was necessary to register each
instance using some `ModelRegistry` magic via the `ModelMapRegistrations`
helper.

This change has simplified the registration of the top-level:
- `ComponentSpecContainer`
- `BinaryContainer`
- `TestSuiteContainer`

Further work will be done to use this for `FunctionalSourceSet`,
which will allow us to remove a bunch of alternative code used for model
registration.