FunctionalSourceSet is a ModelMap subtype
Switch FunctionalSourceSet from a DomainObjectContainer to ModelMap.
This allows us to remove the special-case node initializer for FunctionalSourceSet
and rely on the core SpecializedModelMap infrastructure.

The `AddOnlyRuleAwarePolymorphicDomainObjectContainer` (what a mouthful) is no
longer required.

In order to preserve the existing features of `FunctionalSourceSet` it was
necessary to permit a Managed type with a `ModelMap` property containing
`FunctionalSourceSet` instances.