Revert: Add back projections for reference nodes
We still need to add projections to references, apparently. This
shouldn't be needed, but if there's no actual value referenced, model
report can only show the type of the node if we do this for now. It
should use the schema instead to find the type of the property node
instead.

+review REVIEW-5689