Revert uses of #withType(Jar) back to #getJar
This change partially reverts changes made in commits
d9d68d916fbc507db0ab507919b1ca802924f73a and
2f293ec796f9c65b8f5841d9b85b1f1bb8a1121c, favoring once again the use of
JvmBinaryTasks#getJar over the use of DomainObjectSet#withType. The
reason for the original change was that the class now known as `ApiJar`
originally extended `Jar`, and in doing so broke #getJar's assumptions
about there being one and only one `Jar` for a given binary. Now that
`ApiJar` no longer is-a `Jar`, this assumption is no longer violated,
and thus this commit returns to the original use of #getJar.

Issue: gradle/langos#35 +review REVIEW-5663