Introduce node backed protected BaseBinarySpec.getComponentAs(type)
Binaries are owned by a Component. NativeComponentSpec expose it via its
`getComponent` method. This method is then overriden by subtypes to
specialize its return type.

BaseBinarySpec store a direct reference to some ComponentSpecInternal in
its `owner` field and its `getComponent` method basically returns it.
Specialized versions of this method presents in subtypes simply cast
`owner`. This ComponentSpecInternal happens to be the `privateData` of
the backing node which can't be cast to the corresponding managed public
type (eg. CUnitTSS, GoogleTestTSS).

This commit starts in BaseBinarySpec:

- replace the `owner` field by a `componentNode`
- introduce protected getComponentAs(type) with relies on
  `node.asImmutable(type)` insted of casting

With theses changes, a `BinarySpec` can be owned by a
`@Managed ComponentSpec`.

BaseBinarySpec creation now use a node instead of a
ComponentSpecInternal to refer to its owner component. This explain the
impact on fixtures and tests.

Removed spurious specialized getComponent() methods in NativeBinarySpec
subtypes along the way.

+review REVIEW-5695